refactor: login with token to POST method
